# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Tesis",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/tabla_1_2/get": {
            "get": {
                "tags": [
                    "Table 1 - 2"
                ],
                "summary": "Muestra los Registros",
                "description": "Permite visualizar un listado de todos los registros",
                "operationId": "7d56c9b7665fec03ccc7213033631541",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "campo_1": {
                                            "type": "string",
                                            "example": "Nombre Example"
                                        },
                                        "campo_2": {
                                            "type": "string",
                                            "example": "Apellido Example"
                                        },
                                        "campo_3": {
                                            "type": "string",
                                            "example": "Empresa Example SAC"
                                        },
                                        "state": {
                                            "type": "char",
                                            "example": "A"
                                        },
                                        "table_2": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "campo_1": {
                                                        "type": "integer",
                                                        "example": "+51 975616231"
                                                    },
                                                    "table_1_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "state": {
                                                        "type": "char",
                                                        "example": "A"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No se encuentran los datos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tabla_1_2/create": {
            "post": {
                "tags": [
                    "Table 1 - 2"
                ],
                "summary": "Crea Registros",
                "description": "Crear nuevos Registros para Table 1 y 2",
                "operationId": "b07a972e7e84908b81aca574e24d3d50",
                "requestBody": {
                    "content": {
                        "aplication/json": {
                            "schema": {
                                "properties": {
                                    "campo_1": {
                                        "type": "string"
                                    },
                                    "campo_2": {
                                        "type": "string"
                                    },
                                    "campo_3": {
                                        "type": "string"
                                    },
                                    "table_2": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "campo_1": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "campo_1": "Nombre Example 3",
                                    "campo_2": "Apellido Example 3",
                                    "campo_3": "Empresa Example SAC 3",
                                    "table_2": [
                                        {
                                            "campo_1": "987212211"
                                        },
                                        {
                                            "campo_1": "978131212"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "resp": {
                                            "type": "string",
                                            "example": "Registro creado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "string",
                                            "example": "Error: El Registro no se ha creado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/tabla_1_2/delete/{id}": {
            "delete": {
                "tags": [
                    "Table 1 - 2"
                ],
                "summary": "Inactiva Registros",
                "description": "Inactiva Registros de Table 1 y 2",
                "operationId": "57f426d4cb663bbb88828a6b2f239294",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "resp": {
                                            "type": "string",
                                            "example": "Registro inactivado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error: El registro no se ha inactivado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}